requirements.txt
pylint
pytest
pytest-cov
black
isort
mypy==1.8.0
pyyaml
types-PyYAML